+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Development
 
 * Report configuration modifications of `RSpec.configure` in the final summary.
+* Support conversion of configuration options `backtrace_clean_patterns`, `backtrace_clean_patterns=`, `color_enabled=`, `output`, `output=`, `out`, `out=`, `filename_pattern`, `filename_pattern=`, `backtrace_cleaner`, `backtrace_cleaner=`, `color?` and `warnings`.
 
 ## v2.2.5